[CBRD-26720] Move DB_OOS_STATS to its own client API header
DB_OOS_STATS was defined inline inside dbi.h between extern db_* function
declarations, which broke the convention that dbi.h is a flat list of
externs with type defs imported from dedicated headers.

Move both the struct and the db_oos_stats function declaration to a new
src/compat/db_oos.h, mirroring the db_elo.h pattern. dbi.h now includes
db_oos.h alongside its other db_* type headers, so existing callers
(csql.c via db.h, network_interface_cl.c via db.h) keep working through
the transitive include.

Use int64_t (stdint.h) for recs_sumlen instead of the CUBRID INT64
alias, since db_oos.h is a small standalone client header and doesn't
need to drag in the porting/dbtype_def chain just for one field.
